Remove debounced search
